refac(zk): move `Entity`, `Prover` and `Verifier` to `zk/base/entities`
