+++ tachyon/zk/base/BUILD.bazel
@@ -15,12 +15,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "entity",
-    hdrs = ["entity.h"],
-    deps = ["//tachyon/zk/transcript"],
-)
-
-tachyon_cc_library(
     name = "field_stringifier",
     hdrs = ["field_stringifier.h"],
     deps = [
@@ -34,7 +28,7 @@ tachyon_cc_library(
     hdrs = ["halo2_prover.h"],
     deps = [
         ":halo2_random_field_generator",
-        ":prover",
+        "//tachyon/zk/base/entities:prover",
     ],
 )
 
@@ -85,16 +79,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "prover",
-    hdrs = ["prover.h"],
-    deps = [
-        ":blinded_polynomial",
-        ":blinder",
-        ":verifier",
-    ],
-)
-
-tachyon_cc_library(
     name = "random_field_generator",
     hdrs = ["random_field_generator.h"],
 )
@@ -120,12 +104,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "verifier",
-    hdrs = ["verifier.h"],
-    deps = [":entity"],
-)
-
-tachyon_cc_library(
     name = "verifier_query",
     hdrs = ["verifier_query.h"],
     deps = [":ref"],
@@ -136,16 +114,17 @@ tachyon_cc_unittest(
     srcs = [
         "blinder_unittest.cc",
         "halo2_random_field_generator_unittest.cc",
-        "prover_unittest.cc",
         "ref_unittest.cc",
         "value_unittest.cc",
     ],
     deps = [
-        ":halo2_prover_test",
+        ":blinder",
+        ":halo2_random_field_generator",
         ":ref",
         ":value",
         "//tachyon/base:random",
         "//tachyon/base/containers:container_util",
+        "//tachyon/math/elliptic_curves/bn/bn254:g1",
         "//tachyon/math/finite_fields/test:gf7",
     ],
 )
