+++ tachyon/zk/base/entities/BUILD.bazel
@@ -1,0 +1,31 @@
+load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
+
+package(default_visibility = ["//visibility:public"])
+
+tachyon_cc_library(
+    name = "entity",
+    hdrs = ["entity.h"],
+    deps = ["//tachyon/zk/transcript"],
+)
+
+tachyon_cc_library(
+    name = "prover",
+    hdrs = ["prover.h"],
+    deps = [
+        ":verifier",
+        "//tachyon/zk/base:blinded_polynomial",
+        "//tachyon/zk/base:blinder",
+    ],
+)
+
+tachyon_cc_library(
+    name = "verifier",
+    hdrs = ["verifier.h"],
+    deps = [":entity"],
+)
+
+tachyon_cc_unittest(
+    name = "entities_unittests",
+    srcs = ["prover_unittest.cc"],
+    deps = ["//tachyon/zk/base:halo2_prover_test"],
+)
