+++ tachyon/zk/base/entities/entity.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_BASE_ENTITY_H_
-#define TACHYON_ZK_BASE_ENTITY_H_
+#ifndef TACHYON_ZK_BASE_ENTITIES_ENTITY_H_
+#define TACHYON_ZK_BASE_ENTITIES_ENTITY_H_
 
 #include <memory>
 #include <utility>
@@ -55,4 +55,4 @@ class Entity {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_BASE_ENTITY_H_
+#endif  // TACHYON_ZK_BASE_ENTITIES_ENTITY_H_
