+++ tachyon/zk/plonk/lookup/lookup_argument_runner.h
@@ -11,7 +11,7 @@
 
 #include "gtest/gtest_prod.h"
 
-#include "tachyon/zk/base/prover.h"
+#include "tachyon/zk/base/entities/prover.h"
 #include "tachyon/zk/base/prover_query.h"
 #include "tachyon/zk/plonk/circuit/expressions/evaluator/simple_evaluator.h"
 #include "tachyon/zk/plonk/lookup/lookup_argument.h"
