+++ tachyon/zk/plonk/permutation/grand_product_argument.h
@@ -8,7 +8,7 @@
 
 #include "tachyon/base/parallelize.h"
 #include "tachyon/zk/base/blinded_polynomial.h"
-#include "tachyon/zk/base/prover.h"
+#include "tachyon/zk/base/entities/prover.h"
 
 namespace tachyon::zk {
